% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTempLZ.R
\name{getTempLZ}
\alias{getTempLZ}
\title{Get the observations}
\usage{
getTempLZ(method = NULL, path = NULL, filename = NULL,
  modelTempLZ = NULL)
}
\arguments{
\item{method}{method for processing observations, "lz", "processedNVE"}

\item{path}{Directory where to get the files}

\item{filename}{Filename (path included) of the file to be read}

\item{modelTempLZ}{list(nbLevelZone,Tlr,hfelt,midmett)}
}
\description{
Get and process, if needed, the observations
}
\examples{
\dontrun{
getTempLZ()
}
}
\keyword{tempLZ}

