% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lz.R
\name{lz}
\alias{lz}
\title{Temperature field in to LZ temperature field}
\usage{
lz(temp, modelTempLZ)
}
\arguments{
\item{temp}{temperature field}

\item{modelTempLZ}{list(nbLevelZone,Tlr,hfelt,midmett)}
}
\description{
Temperature field in to LZ temperature field
}
\examples{
\dontrun{
lz()
}
}
\keyword{tempLZ}

