% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\title{Temperature field in to LZ temperature field}
\usage{
perturbation(temp, modelPert)
}
\arguments{
\item{temp}{temperature field}

\item{modelPert}{list(ditr,param)}
}
\description{
Temperature field in to LZ temperature field
}
\examples{
\dontrun{
perturbation()
}
}
\keyword{tempLZ}

