% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dddTempLZ.stateX.R
\name{stateX}
\alias{stateX}
\title{Temperature for each elevation zone}
\usage{
stateX(v, modelTempLZ)
}
\arguments{
\item{v}{temperature value (scalar)}

\item{modelTempLZ}{list(nbLevelZone=a,Tlr=b,hfelt=c,midmett=d)}
}
\description{
The function \code{stateX()} process the temperature for each elevation zone
}
\examples{
\dontrun{
dddTempLZ.stateX()
}
}
\keyword{TempLZ}

