% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTempLZ.R
\name{getTempLZ}
\alias{getTempLZ}
\title{Get the observations}
\usage{
getTempLZ(method = NULL, path = NULL, filename = NULL, ...)
}
\arguments{
\item{method}{method for processing observations, "lz", "processedNVE"}

\item{path}{Directory where to get the files}

\item{filename}{Filename (path included) of the file to be read}

\item{...}{options for the different methods}
}
\description{
Get and process, if needed, the observations
}
\examples{
\dontrun{
getTempLZ()
}
}
\keyword{tempLZ}

