% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.R
\name{stateX}
\alias{stateX}
\title{Temperature for each elevation zone}
\usage{
stateX(v, modelTempLZ)
}
\arguments{
\item{v}{temperature value (scalar)}

\item{modelTempLZ}{list(nbLevelZone,Tlr,hfelt,midmett)}
}
\description{
The function \code{stateX()} process the temperature for each elevation zone
}
\examples{
\dontrun{
stateX()
}
}
\keyword{TempLZ}

